Wrote Timer support.
SEEN8502 (Last Regrets) plays all the way through...but brokenly.
There is no sound. The lack of ANM file leads to the weird glitches
where instead of fading in the CG of a character ontop of their inked
sketch, we get a flash of the tree.

There is also weirdness with "Yakusoku, dayo..." being displayed in the
top left corner of the screen when it really shouldn't be.