Finally finally finally get subtractive filtering working on the text box. I've lost half a week to this.
And the code is basically one big hack that requires newer video
cards: Since I couldn't figure out how to express the operation as a
set of glBlend statements (and I think it may not be possible), I use
a fragment shader with the proper equation. I'm going to have to come
back and write some graceful degradation code.