Fix GraphicsObject's operator=() method.
When copying a GraphicsObject on top of an old GraphicsObject, make
sure that if the incoming object has null m_objectData NULL, we set
the current object's m_objectData to NULL too. 

This patch fixes the problem where in CLANNAD, during the entire "kono
gakou ha, suki desu ka" opening scene, saving and loading would
display the blue sky object from the previous scene, in addition to a
piece of the credits from even earlier. (Because they weren't properly
overwritten with a NULL)