Assume that textouts reset the graphical update mode
I assume this is what is going on after a few tests I did with the
RealLive interpreter; when text was displayed, it changed the
behaviour of a recLoad() to automatic display. I am not fully sure if
it was text being displayed on the screen which changed the update
mode or the start of text changing the mode (as I've implemented it).

This fixes the problems with selecting the third bookmark in
Planetarian since I made the rest of the system care about the update
mode starting in 7b4a31767bd30d3043cb22c00f9488158d405496.

TextoutLongOperation may be the wrong place to put this; later on I
may want to add this somewhere in the TextSystem.