Smooth scrolling while DrawManual() is active.
While DrawManual() is active, we can't redraw the screen
when the cursor moves. So instead of performing a full
SDLGraphicsSystem::refresh(), we take the current contents
of the frame buffer, save where we're going to display the
cursor, render the cursor, SDL_GL_SwapBuffers() and then
restore what the cursor drew over so we can repeat this
process the next time the user moves the mouse.

This works because we only glClear when we start rendering
a full new frame.