Keep track of whether the mouse hotspot is inside the window.
We don't want to redraw the custom cursor when the mouse is
outside the window so when the mouse leaves the window, force
a redraw.

Also, don't respond to mouse move events and clicks when the
hotspot is outside the window. OSX will continue to throw
mouse events at you...