Make mouse cursor handling work under osx.
Bug: Switch out of osx to another app. Switch back. Mouse
cursor is missing.

Fix: Respond to the SDL_APPINPUTFOCUS event, which we get when
we received keyboard focus. This is different from X11; X11 windows
seem to get mouse events as long as the mouse cursor is over a
visible portion of the window, so this doesn't happen under linux...