Fix animations not playing in ganPlayEx when DrawManual is on.
Fix bug introduced in 7b4a31767bd30d3043cb22c00f9488158d405496. I
revamped how the screen was being updated, but didn't consider the case
of when we were in DrawManual() mode and

Solve the problem by making saving the screen update mode and setting to
automatic when entering WaitForGanToFinish, and restore the old value
when we exit.