Force a sleep when refresh() is called when in DrawManual() mode.
Spinning on pieces of byte-code usually occur only when we're in
DrawManual() mode. We can't just sleep after every command, so this is
about the best we can do to reduce CPU usage in these cases.

Cuts CPU usage from about %70 to %40 on my old Pentium M during the Kanon
ED sequence.