Redo how events are passed around and remove the EventHandler interface.
This removes the EventHandler interface and all management of which
EventHandler gets which events, which has constantly been wrong. All
LongOperations are now EventListeners and events are only delivered to
any LongOperation when that LongOperation is on the top of the callstack.
(Instead of trying to keep a separate stack in EventSystem synchronized
with the RLMachine callstack like I was trying to do before...)