First semi-working implementation of rlBabel.
Copies most of rlBabel's texteout.cpp into rlvm's RlBabelDLL file. The
code is very similar but not exact since I copy-pasted it and then
modified it to use rlvm's internal structure, and then did a round of
clean ups to make things more readable and to use C++ constructs instead
of raw arrays of chars.