Fix problem with Kanon opening never showing.
This is a revert of 05899008696c384718da8fc911607bb08160e2a3, which I
think was the wrong solution. I'm not sure what problem I was talking
about with clicks not registering, but a quick test on the English Kanon
patch shows no problems, possibly because the correct fix was the pause
behaviour fixed in 4026b2c9642a37c712c661829abfc5090c2234b1.