Fix goto_on() cases not being parsed properly during bytecode tokenization.
SEEN2416 of the CLANNAD_FV english patch would crash as soon as it was
farcalled(). The problem was that goto_on was not being tokenized correctly.
GotoOnElement::length() was implemented in terms of a vector named |targets|
which wouldn't have been filled with its data at the time
GotoOnElement::length() would be called during tokenization. The
GotoOnElement filled in the |target_ids| vector, and filled |targets| from
|target_ids| in a step that came latter.

Instead, uses argc(), which should equals target_ids.size() and
targets.size().