Implement cos().
The unexpectedly fixes a lot of the ALMA opening, including the second
background image for each character during the OP, and everything
about the final part of the OP with the text messages.

The Bonbee! people must really love trig because I don't see the
relevancy since the speed of movement there looks linear...