Fix Aimi's epilogue by removing false positive error.
P_BRIDE, SEEN2214, Line 1343 appears to have an error in it.

It starts with: "\　\{恭一}"

The first 'escaped' space shouldn't be there and was causing a false
positive on error checking code I had in TextWindow::display_char()
which made sure names had been handled by the time
TextWindow::display_char() was called.