Clears the graphics stack during openBg commands.
Confirmed (in P_BRIDE) that at least some openBg commands cleared the
stack. This speeds loading of save files from O(n) (where n is the number of
graphics displayed on the screen *ever* in the current game session) to O(1)
(where the constant is s small number of graphics commands since the last
openBg).