Revert "DCs should start out transparent."
This reverts commit 4656a3803c41cee619458af7be464d3c0d3eb615.

Further experimentation shows this to be questionable; while there
can be transparent DCs, it doesn't appear to matter because it looks
like either bgr OR dc0 is used as the background; never both.

The blitting parts are likewise wrong because we don't want to blend
dc1 with dc0--we want to replace it even if we blended a source image
onto dc1.