Do a better job of implementing the Japanese line breaking rules.
We now squeeze one last piece of punctuation onto a line instead of
breaking and we will break for multiple pieces of punctuation.

Tested by comparing (line by line) the first chapter of Planetarian
and the CLANNAD Prologue. If there's any weirdness with the text
system in the near future, this is the patch to suspect.