Fix the graphics stack in most games.
Previously, we used the dedicated GraphicsStackFrame class to keep track
of commands that were in the graphics stack, requiring manual
registration in each implementation of a command that goes on the stack,
and requires manual interpretation to rerun the commands.

The new implementation creates a list of RealLive bytecode strings which
represent the commands and can use the normal command dispatch mechanism.
Opcode implementations are wrapped automatically with MappedRLModule.

This also fixes a few smaller fixes, including tracking the default grp
names correctly.