Adds a script for the Refrain/Little Busters chapters.
This isn't fully automated. It doesn't drive the main menu, which is harder
since it may change the main menu implementation, and it needs to inject
several clicks half way through.

Still, this script (with some human intervention) clears the final chapter of
Little Busters, which is now fully supported.