Memory saved to disk should be memory at the state of the last Savepoint().
The weird behaviour with restoring a game where the current line has a gloss
comes from rlBabel writing to the string variable that's used to strout()
while waiting for a click to match against the gloss region. If we hit a
Savepoint(), we should later save the memory at the time of that Savepoint()
to disk.

Keep a list of original values next to the main memory bank, copying original
data to the list before modifying it. We can just clear this list at the next
Savepoint(), and can use it to recreate the data we need to save to disk in
that less common case.

Also adds unit tests to exercise this complex new behaviour.

Closes #3