Implement raw values ({5}) in index_series.
Raw values act as if they change the current value if the previous term in the
index_series has finished. I think this is a giant hack, but it produces the
correct results, which is making the clouds scenes in Little Busters sync the
multiple bitmaps used in that scene correctly.