Reimplement findFile().
Profiling showed that 10% of runtime was spent in findFile(), which searched
the filesystem over and over for each listed directory/extension
combination. In retrospect, it was pathologically bad.

Implement a new version which iterates over the directory tree of the game
first, lower cases path components so we don't do the expensive correctCase(),
and then returns from this cache instead of going to disk (and doing the
expensive directory iteration on each try).

Closes GH-9