Add support for reading and usage of tone curve data from TCC data
While much of the TCC file "header" ended up being ignored during
reverse engineering and the way the file is read therefore appears
rather crude, it appears as if all the data that needs to be read can
be read using this simplified method (all of the reading routines are
in ToneCurve.cpp).  To put it simply, TCC files are made up of 3 sets
of 256 unsigned chars per effect where the 3 sets represent the R, G,
and B channels and the 256 unsigned chars represent the output color
value under the effect for each input color value, where the input
color value acts as the index.  This has only been tested on Memorial
Edition games due to lack of availability of other games (and Linux due
to difficulties setting up a development toolchain on my Mac machine),
but it has been tested on games that have differing amounts of tone
curve effects and appears to give the desired results, and TCC files
from other versions of RealLive (mostly newer) have been examined and
manually tested and also appear to work.  The tone curve effects were
implemented as colour transformations as this appears to be the most
natural way to do it - as a result it was necessary to fix
TransformSurface to accomodate surfaces with alpha.  Hopefully this
stands up to more comprehensive testing!