Smooth animations by uploading objects outside of timing loops.
In rec/grp heavy games, the correct thing is *not* to upload loaded images to
the GPU, because they'll be blitted to one of the DCs. But when using objects,
we want to make sure that the data is on the graphics card by the time the load
command completes because otherwise there will be a noticeable lag in any
animation that uses the object.