Implement obj{Light,Tint,Colour} in a fragment shader.
This fixes parts of the Little Busters refrain, among others. The ending
reprise of Little Busters shows them driving in the bus for the first
time. This is implemented as a progressive enhancement and the old path is the
fallback if there's no fragment shader support.

This also cleans up all fragment shader code and puts them in their own file.