Fix parsing g00 files for the LB Refrain ending.
Several g00 images in the ending are stacked right on top of each other,
leading to corruption. (For example, the credits and the page flipping in the
Refrain.)  After a few weeks of trying to do this the right way, I'm giving up
and just hacking this. I've tried looking at all the data in the g00 file and I
haven't found anything that looks like a layer flag. After grepping all the g00
files I have access to, I've verified that this condition only happens in LB
and in 4 files in Tomoyo After, which I suspect I wasn't displaying correctly
either.

Closes GH-39