Fix color order on newer SDL + Intel Macbook Airs
Previously, we were converting some surfaces to the display format.
This appears to do nothing on Linux, but on some recent combination of
OSX 10.8 / newer SDL / the Intel integrated graphics in my new Macbook
AIr, this actually does convert the surface to a different format,
which breaks a whole bunch of rendering.