Directory renaming part [01/12]: src/effects/
I've made multiple stabs in the past to cleanup the rlvm code to be in
line with the Chromium style guide. This time around, I'm making sure
the filenames follow the style guide. This means all file name
components are lowercased, using unix_hacker style instead of CamelCase,
and use cc/h instead of cpp/hpp.