Start fixing up the include order.
Order up *most* of the includes according to Google style. This still
isn't enough to remove the cpplint build/include_order suppression. The
boost::archive headers and glew.h must come before any of the rest of
the headers. The other half of fixing this is making sure that NOLINT
works on #include lines, which it doesn't yet.