pause() should start new pages on all open windows.
Previously, we were only starting a new page on pause R_COMMAND_MOD=0
mode. We should have been starting new pages on all open windows
instead. This is the root cause behind not clearing the main text window
when the user clicks on a gloss which overflows the gloss window.

Closes GH-5.