Don't snapshot empty pages for scrollback.
In the English Kanon patch, if you click a gloss, and then go forward a
few lines, and then look through scrollback, you'd have seen an empty
"page" that didn't have a text window. Add a check during snapshot where
we won't add an empty TextPage to the list of snapshots.