Add rudimentary Kud Wafter support.
Currently, the game appears to start up, and it seems to be playable,
but we get a CGM error on startup. It looks like the CGM file has
another xor key on it or something. Will require further investigation;
CG progress is probably broken. However, basically everything else looks
like it works.

Thanks @jyuichi!