Fix Western, monospace text output.
When not using rlbabel, we always measure characters in terms of
MOJI_SIZE instead of the actual font's metrics. Notably, several
Japanese fonts on Linux have completely busted minx/advance metrics for
western characters, and should not be consulted.

The current text breaking rule appears to be: break a line if I can't
fit another fullwidth character unless the next character is a
space. This also removes a block which was inserting a fullwidth space
at the beginning of Yumemi's lines. (This may be controlled by a
Gameexe.ini key though; when I have more time, I should double check
that this doesn't introduce a regression in Japanese planetarian.)

This fixes text output on at least the first 30 screens of text in the
English planetarian.

(This also fixes Japanese Clannad, which was putting one less character
on lines than it should have been.)