Dont lazily parse the ExpressionElement.
Instead of storing the raw bytecode string of expression elements
and then parsing them on demand, just parse the expression now
that the ExpressionPiece is cheap.

This similarly changes parts of the selection interface. (Also,
I don't see any uses of passing a non-default window number to
select_w in any games that I've disassembled locally.)