Fix audio quality in SDL.
This is a cherry-pick of da0ccd73819c081b96cbfae61ea50503f9011d16 from
the SDL2 branch; the fix also works in old SDL.

This forces the audio quality to 44.1 khz audio. This fixes the static
that showed up on some sounds in the sdll2 port. (When applying this
patch to the old sdl1 branch, it also fixes the problem where some of
the sound effects are higher pitched than they should be.)

Conflicts:
	src/systems/sdl/sdl_sound_system.cc