End TextoutLongOperation on mouse up, not mouse down.
Previously, when the user would push the mouse button down, that event
was routed to the TextoutLongOperation, which would replaced itself with
the PauseLongOperation. That pause operation then received the mouse up
event and would immediately continue. By making TextoutLongOperation
only act on mouse up, we don't leak a mouse up to the next long
operation.

(Tested in the English version of planetarian.)

Closes #75, #59.