Extend the stacked g00 parsing code.
In 28e4a60, we started dealing with a sort of g00 file that was first
used in Tomoyo After, where all regions had the same boundaries. This
previously broke the Little Busters Refrain, and now is causing problems
for the waku buttons in the Clannad English Edition.

This extends the logic so that it detects images that have only one
region specified multiple times. It looks like the logic which allocates
more room doesn't need to change.

Closes GH-69.